fix: refactor generate attendance image functions
